+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ This project adheres to `Semantic Versioning`_ starting with version 0.11.0.
 Added
 -----
 - add formatter 'black'
+- Slots filled before the start of a form are now validated upon form start
+- In debug mode, the values of required slots for a form are now printed
+  before submitting
 
 Changed
 -------